remove exception for unknown postal code formats
in practical terms, raising an exception in this case is not useful
when trying to validate the postal code of a model. since the list
of known formats is not comprehensive, it's totally reasonable for
a globally-used application to encounter countries that don't have
known postal code formats. so when that happens, just accept
whatever value was given.

also add postal code formats for GB, NZ and SG.