Fix NINO array literal syntax
The commas were included in the generated array elements, so NINOs were
never rejected for having unallocated or administrative prefixes. An
example number that should have been rejected but wasn't is GB123456C.

Unfortunately, the tests didn't catch this mistake because they used the
same incorrect syntax as the implementation. This commit fixes both the
implementation and the tests.